requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
babel==2.12.1 ; python_version >= "3.11" and python_version < "4.0"
django-money==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.11" and python_version < "4.0"
py-moneyed==3.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
django-admin-rangefilter==0.11.2
django-bootstrap3==23.4
django-filter==23.3
django-tables2==2.6.0